new component functions directory
