+++ smooth/components/component_functions/all_component_functions.py
@@ -1,0 +1,25 @@
+import numpy as np
+from scipy import interpolate
+
+
+# Function for CompressorH2 component in order to calculate compressibility factor
+def calculate_compressibility_factor(p_in, p_out, temp_in, temp_out):
+    temp = np.transpose([200, 300, 400, 500, 600, 800, 1000, 2000])
+
+    p = [1, 10, 20, 40, 60, 80, 100, 200, 400, 600, 800, 1000]
+
+    z = [[1.0007, 1.0066, 1.0134, 1.0275, 1.0422, 1.0575, 1.0734, 1.163, 1.355, 1.555, 1.753, 1.936],
+         [1.0005, 1.0059, 1.0117, 1.0236, 1.0357, 1.0479, 1.0603, 1.124, 1.253, 1.383, 1.510, 1.636],
+         [1.0004, 1.0048, 1.0096, 1.0192, 1.0289, 1.0386, 1.0484, 1.098, 1.196, 1.293, 1.388, 1.481],
+         [1.0004, 1.0040, 1.0080, 1.0160, 1.0240, 1.0320, 1.0400, 1.080, 1.159, 1.236, 1.311, 1.385],
+         [1.0003, 1.0034, 1.0068, 1.0136, 1.0204, 1.0272, 1.0340, 1.068, 1.133, 1.197, 1.259, 1.320],
+         [1.0002, 1.0026, 1.0052, 1.0104, 1.0156, 1.0208, 1.0259, 1.051, 1.100, 1.147, 1.193, 1.237],
+         [1.0002, 1.0021, 1.0042, 1.0084, 1.0126, 1.0168, 1.0209, 1.041, 1.080, 1.117, 1.153, 1.187],
+         [1.0009, 1.0013, 1.0023, 1.0044, 1.0065, 1.0086, 1.0107, 1.021, 1.040, 1.057, 1.073, 1.088]]
+
+    interp_func = interpolate.interp2d(p, temp, z)
+
+    z_in = interp_func(p_in, temp_in)
+    z_out = interp_func(p_out, temp_out)
+
+    return [z_in, z_out]
